@@ -31,9 +31,9 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.1",
         "ext-json": "*",
-        "php-http/discovery": "^1.7",
+        "php-http/discovery": "^1.19",
         "psr/http-client": "^1.0",
         "symfony/polyfill-php81": "^1.33"
     },
@@ -54,7 +54,7 @@
         "symfony/http-client": "Use Symfony Http client"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5 || ^10.5",
+        "phpunit/phpunit": "^10.5",
         "php-cs-fixer/shim": "^3.59.3",
         "http-interop/http-factory-guzzle": "^1.2.0",
         "phpstan/phpstan": "^2.0",
