@@ -46,7 +46,7 @@ class Client implements Http
         ?ClientInterface $httpClient = null,
         ?RequestFactoryInterface $reqFactory = null,
         array $clientAgents = [],
-        ?StreamFactoryInterface $streamFactory = null
+        ?StreamFactoryInterface $streamFactory = null,
     ) {
         $this->baseUrl = $url;
         $this->http = $httpClient ?? Psr18ClientDiscovery::find();
